backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
python-dotenv==1.0.1
pytest==8.3.3
pylint==3.3.1
black==24.10.0
gunicorn==23.0.0